feat(Astro): Introduce new Footer and Header components, refactor layout structure, and remove deprecated files
